Move authentication related components
