+++ src/components/Authentication/index.js
@@ -1,0 +1,5 @@
+import Login from './Login'
+import Logout from './Logout'
+import SignUp from './SignUp'
+
+export { Login, Logout, SignUp }
